[LEMS-4203/linear-goes-oprah-refactor] [Interactive Graph] Refactor movable-line
