+++ packages/perseus/src/widgets/interactive-graphs/graphs/components/movable-line/index.ts
@@ -1,0 +1,6 @@
+export {MovableLine} from "./movable-line";
+export {
+    getMovableLineKeyboardConstraint,
+    insetTipAlongRay,
+    trimRange,
+} from "./util";
